Construct execution job vertex, execution job edge, execution vertex and execution edge in one method in graph manager.

+++ streaming/java/streaming-runtime/src/main/java/org/ray/streaming/runtime/core/transfer/ChannelID.java
@@ -154,10 +154,10 @@ public class ChannelID {
       |    Head    | Timestamp | Empty | From  |  To    |
       | 8 bytes    |  4bytes   | 4bytes| 2bytes| 2bytes |
     */
-    Preconditions.checkArgument(fromTaskId < Short.MAX_VALUE,
-        "fromTaskId %d is larger than %d", fromTaskId, Short.MAX_VALUE);
-    Preconditions.checkArgument(toTaskId < Short.MAX_VALUE,
-        "toTaskId %d is larger than %d", fromTaskId, Short.MAX_VALUE);
+    Preconditions.checkArgument(fromTaskId < Integer.MAX_VALUE,
+        "fromTaskId %d is larger than %d", fromTaskId, Integer.MAX_VALUE);
+    Preconditions.checkArgument(toTaskId < Integer.MAX_VALUE,
+        "toTaskId %d is larger than %d", fromTaskId, Integer.MAX_VALUE);
     byte[] channelName = new byte[20];
 
     for (int i = 11; i >= 8; i--) {
